rename leader_svc
